examples/load_artifacts/requirements.txt
pillow >= 9.3 
